feat(ui): move file browser toggle into content headers
Relocate the file browser toggle out of the global title bar into the
project canvas header and the session modal header, placed before the
project name so it sits next to the panel it controls.
